// frontend/src/types/events.d.ts
type Event = {
	id: number
	title: string
	date: Date
	scheduleTime: string
	image: string
	description: string
	location: string
	locationMap: string
	locationDescription: string
	price: number
	priceFinal: number
	release1: number
	release2: number
	release3: number
	priceVIP: number
}

export type { Event }